Split oversized backup.js and members.js routes (ImprovementPlan Chunk 9, M1)
Refactor the two largest backend route files into sub-router directories
following the finance/ precedent. Pure moves, no behaviour change.

- routes/backup.js (2,353 lines) -> routes/backup/: export.js (export route +
  sheet builders), restore.js (restore helpers used by system.js), shared
  helpers.js (str), index.js (requireAuth + mount + re-export restore helpers).
- routes/members.js (1,970 lines) -> routes/members/: list.js (listings/
  statistics/validate/download), lifecycle.js (renewals/renew/non-renewals/
  lapse), crud.js (single-member CRUD/photo/groups), shared helpers.js
  (resolveGiftAidAmount, deriveInitials), index.js (mounts literal-path routers
  before the /:id CRUD router to preserve route match order).
- Updated import sites: app.js, system.js, restoreBeacon.test.js.

Route registrations verified identical before/after. Backend suite green
(45 files, 593 tests); lint + Prettier clean.
