+++ backend/src/app.js
@@ -18,7 +18,7 @@ import privilegeRoutes from './routes/privileges.js';
 import systemRoutes from './routes/system.js';
 import memberClassRoutes from './routes/memberClasses.js';
 import memberStatusRoutes from './routes/memberStatuses.js';
-import memberRoutes from './routes/members.js';
+import memberRoutes from './routes/members/index.js';
 import facultyRoutes from './routes/faculties.js';
 import venueRoutes from './routes/venues.js';
 import groupRoutes from './routes/groups.js';
@@ -28,7 +28,7 @@ import financeRoutes from './routes/finance/index.js';
 import pollRoutes from './routes/polls.js';
 import auditRoutes from './routes/audit.js';
 import officeRoutes from './routes/offices.js';
-import backupRoutes from './routes/backup.js';
+import backupRoutes from './routes/backup/index.js';
 import addressExportRoutes from './routes/addressExport.js';
 import emailRoutes from './routes/email.js';
 import giftAidRoutes from './routes/giftAid.js';
