+++ backend/src/routes/members/crud.js
@@ -1,952 +1,19 @@
-// beacon2/backend/src/routes/members.js
+// beacon2/backend/src/routes/members/crud.js
+// Single-member CRUD: fetch, create, update, delete, photo, and group membership.
 
 import { Router } from 'express';
 import { randomBytes } from 'crypto';
 import { z } from 'zod';
-import ExcelJS from 'exceljs';
-import PDFDocument from 'pdfkit';
-import { requireAuth } from '../middleware/auth.js';
-import { requirePrivilege } from '../middleware/requirePrivilege.js';
-import { tenantQuery, escapeLike } from '../utils/db.js';
-import { hashOpaqueToken } from '../utils/password.js';
-import { sanitizeCell } from '../utils/spreadsheet.js';
-import { AppError } from '../middleware/errorHandler.js';
-import { isFeatureEnabled, requireFeature } from '../middleware/requireFeature.js';
-import { logAudit } from '../utils/audit.js';
-import { decodeAndValidateImage } from '../utils/uploads.js';
+import { requirePrivilege } from '../../middleware/requirePrivilege.js';
+import { requireFeature } from '../../middleware/requireFeature.js';
+import { tenantQuery } from '../../utils/db.js';
+import { hashOpaqueToken } from '../../utils/password.js';
+import { AppError } from '../../middleware/errorHandler.js';
+import { logAudit } from '../../utils/audit.js';
+import { decodeAndValidateImage } from '../../utils/uploads.js';
+import { resolveGiftAidAmount, deriveInitials } from './helpers.js';
 
 const router = Router();
-router.use(requireAuth);
-
-// ─── helpers ──────────────────────────────────────────────────────────────
-
-/** Look up the Gift Aid eligible amount for a member at a given date.
- *  Returns null if GA is not enabled, member has no GA declaration, or no GA fee configured.
- */
-async function resolveGiftAidAmount(slug, memberId, classId, transactionDate) {
-  // Check if GA feature is enabled
-  if (!(await isFeatureEnabled(slug, 'giftAid'))) return null;
-
-  const [settings] = await tenantQuery(
-    slug,
-    `SELECT fee_variation FROM tenant_settings WHERE id = 'singleton'`,
-  );
-
-  // Check if member has a GA declaration at or before the transaction date
-  const [m] = await tenantQuery(slug, `SELECT gift_aid_from FROM members WHERE id = $1`, [
-    memberId,
-  ]);
-  if (!m?.gift_aid_from) return null;
-  const gaFrom = String(m.gift_aid_from).slice(0, 10);
-  const txDate = String(transactionDate).slice(0, 10);
-  if (gaFrom > txDate) return null;
-
-  // Look up the GA fee from the class
-  if (settings.fee_variation === 'varies_by_month') {
-    // Month-specific: find the fee for the transaction month
-    const txMonth = new Date(txDate).getMonth() + 1; // 1-12
-    const [monthFee] = await tenantQuery(
-      slug,
-      `SELECT gift_aid_fee::float FROM class_monthly_fees WHERE class_id = $1 AND month_index = $2`,
-      [classId, txMonth],
-    );
-    return monthFee?.gift_aid_fee ?? null;
-  }
-
-  // Standard: use the class-level gift_aid_fee
-  const [cls] = await tenantQuery(
-    slug,
-    `SELECT gift_aid_fee::float FROM member_classes WHERE id = $1`,
-    [classId],
-  );
-  return cls?.gift_aid_fee ?? null;
-}
-
-/** Derive initials from a forenames string: "William John" → "WJ" */
-function deriveInitials(forenames) {
-  return (forenames ?? '')
-    .split(/\s+/)
-    .filter(Boolean)
-    .map((n) => n[0].toUpperCase())
-    .join('');
-}
-
-// ─── GET /members ─────────────────────────────────────────────────────────
-// Query params:
-//   status      – comma-separated list of status IDs  (default: all)
-//   classId     – single class ID
-//   pollId      – filter to members in this poll
-//   negatePoll  – '1' to invert: members NOT in the poll
-//   q           – free-text search
-//   letter      – single letter to filter surname start
-//   paymentMethod – filter by last membership transaction's payment method
-
-router.get('/', requirePrivilege('members_list', 'view'), async (req, res, next) => {
-  try {
-    const slug = req.user.tenantSlug;
-    const { status, classId, pollId, negatePoll, q, letter, cf, paymentMethod } = req.query;
-
-    const conditions = [];
-    const params = [];
-    let i = 1;
-
-    if (status) {
-      const ids = status.split(',').filter(Boolean);
-      if (ids.length) {
-        conditions.push(`m.status_id = ANY($${i++}::text[])`);
-        params.push(ids);
-      }
-    }
-
-    if (classId) {
-      conditions.push(`m.class_id = $${i++}`);
-      params.push(classId);
-    }
-
-    if (pollId) {
-      if (negatePoll === '1') {
-        conditions.push(`m.id NOT IN (SELECT member_id FROM poll_members WHERE poll_id = $${i++})`);
-      } else {
-        conditions.push(`m.id IN (SELECT member_id FROM poll_members WHERE poll_id = $${i++})`);
-      }
-      params.push(pollId);
-    }
-
-    if (letter && /^[A-Z]$/i.test(letter)) {
-      conditions.push(`upper(m.surname) LIKE $${i++}`);
-      params.push(letter.toUpperCase() + '%');
-    }
-
-    if (q) {
-      const like = `%${escapeLike(q)}%`;
-      conditions.push(`(
-        m.surname        ILIKE $${i}   OR
-        m.forenames      ILIKE $${i}   OR
-        m.known_as       ILIKE $${i}   OR
-        m.email          ILIKE $${i}   OR
-        m.mobile         ILIKE $${i}   OR
-        a.street         ILIKE $${i}   OR
-        a.town           ILIKE $${i}   OR
-        a.postcode       ILIKE $${i}   OR
-        m.custom_field_1 ILIKE $${i}   OR
-        m.custom_field_2 ILIKE $${i}   OR
-        m.custom_field_3 ILIKE $${i}   OR
-        m.custom_field_4 ILIKE $${i}   OR
-        m.membership_number::text = $${i + 1}
-      )`);
-      params.push(like, q.trim());
-      i += 2;
-    }
-
-    if (cf) {
-      const cfLike = `%${escapeLike(cf)}%`;
-      conditions.push(`(
-        m.custom_field_1 ILIKE $${i}   OR
-        m.custom_field_2 ILIKE $${i}   OR
-        m.custom_field_3 ILIKE $${i}   OR
-        m.custom_field_4 ILIKE $${i}
-      )`);
-      params.push(cfLike);
-      i += 1;
-    }
-
-    if (paymentMethod) {
-      conditions.push(`m.id IN (
-        SELECT lt.member_id_1
-        FROM (
-          SELECT DISTINCT ON (t.member_id_1) t.member_id_1, t.payment_method
-          FROM transactions t
-          WHERE t.member_id_1 IS NOT NULL AND t.type = 'in'
-          ORDER BY t.member_id_1, t.date DESC, t.created_at DESC
-        ) lt
-        WHERE lt.payment_method = $${i++}
-      )`);
-      params.push(paymentMethod);
-    }
-
-    const where = conditions.length ? `WHERE ${conditions.join(' AND ')}` : '';
-
-    const members = await tenantQuery(
-      slug,
-      `SELECT m.id, m.membership_number, m.title, m.forenames, m.surname,
-              m.known_as, m.email, m.mobile, m.hide_contact,
-              ms.id AS status_id, ms.name AS status,
-              mc.id AS class_id,  mc.name AS class,
-              a.house_no, a.street, a.town, a.postcode, a.telephone,
-              m.joined_on, m.next_renewal, m.partner_id,
-              m.custom_field_1, m.custom_field_2,
-              m.custom_field_3, m.custom_field_4,
-              (m.portal_password_hash IS NOT NULL) AS has_portal_password,
-              m.portal_email_verified
-       FROM members m
-       LEFT JOIN member_statuses ms ON ms.id = m.status_id
-       LEFT JOIN member_classes  mc ON mc.id = m.class_id
-       LEFT JOIN addresses        a ON a.id  = m.address_id
-       ${where}
-       ORDER BY m.surname, m.forenames`,
-      params,
-    );
-
-    res.json(members);
-  } catch (err) {
-    next(err);
-  }
-});
-
-// ─── GET /members/validate ────────────────────────────────────────────────
-// Returns all members with their address data for client-side data quality checks.
-// Requires settings:view (admin only).
-
-// ─── GET /members/recent ──────────────────────────────────────────────────
-// Returns members who joined in the given date range.
-// Query params: from (ISO date), to (ISO date). Defaults: last 30 days.
-
-router.get('/recent', requirePrivilege('members_recent', 'view'), async (req, res, next) => {
-  try {
-    const slug = req.user.tenantSlug;
-    const toDate = req.query.to || new Date().toISOString().slice(0, 10);
-    const fromDate =
-      req.query.from ||
-      (() => {
-        const d = new Date();
-        d.setDate(d.getDate() - 30);
-        return d.toISOString().slice(0, 10);
-      })();
-
-    const rows = await tenantQuery(
-      slug,
-      `SELECT m.id, m.membership_number, m.forenames, m.surname, m.known_as,
-              m.email, m.mobile, a.telephone, m.joined_on, m.next_renewal,
-              mc.name AS class_name,
-              ms.name AS status_name,
-              a.house_no, a.street, a.town, a.postcode,
-              m.portal_password_hash IS NOT NULL AS has_portal_password,
-              m.portal_email_verified
-       FROM members m
-       LEFT JOIN member_classes   mc ON mc.id = m.class_id
-       LEFT JOIN member_statuses  ms ON ms.id = m.status_id
-       LEFT JOIN addresses         a ON a.id  = m.address_id
-       WHERE m.joined_on >= $1::date
-         AND m.joined_on <= $2::date
-       ORDER BY m.joined_on DESC, m.surname, m.forenames`,
-      [fromDate, toDate],
-    );
-    res.json(rows);
-  } catch (err) {
-    next(err);
-  }
-});
-
-// ─── GET /members/statistics ──────────────────────────────────────────────
-// Returns membership and group statistics.
-// Query params: from, to — date range for section 4 (Members by Renew Date).
-
-router.get(
-  '/statistics',
-  requirePrivilege('membership_statistics', 'view'),
-  async (req, res, next) => {
-    try {
-      const slug = req.user.tenantSlug;
-
-      // Fetch settings for year start and renewal periods
-      const [cfg] = await tenantQuery(
-        slug,
-        `SELECT year_start_month, year_start_day, advance_renewals_weeks, grace_lapse_weeks
-       FROM tenant_settings WHERE id = 'singleton'`,
-      );
-      const yearStartMonth = cfg?.year_start_month ?? 1;
-      const yearStartDay = cfg?.year_start_day ?? 1;
-      const advanceWeeks = cfg?.advance_renewals_weeks ?? 4;
-      const graceLapse = cfg?.grace_lapse_weeks ?? 4;
-
-      // Compute current membership year start date
-      const now = new Date();
-      const thisYear = now.getFullYear();
-      const candidateStart = new Date(thisYear, yearStartMonth - 1, yearStartDay);
-      const yearStartDate =
-        candidateStart <= now
-          ? candidateStart
-          : new Date(thisYear - 1, yearStartMonth - 1, yearStartDay);
-      const yearStartIso = yearStartDate.toISOString().slice(0, 10);
-
-      // Section 2: Current members by class
-      const classStats = await tenantQuery(
-        slug,
-        `SELECT mc.id, mc.name,
-              COUNT(m.id)::int                                    AS total,
-              COUNT(m.id) FILTER (WHERE m.email IS NOT NULL AND m.email <> '')::int AS with_email,
-              COUNT(m.id) FILTER (WHERE m.joined_on >= $1::date)::int               AS first_year,
-              COUNT(m.id) FILTER (WHERE m.joined_on <  $1::date)::int               AS second_year_plus
-       FROM member_classes mc
-       LEFT JOIN members m ON m.class_id = mc.id
-                           AND m.status_id IN (
-                               SELECT id FROM member_statuses WHERE name ILIKE '%Current%'
-                           )
-       GROUP BY mc.id, mc.name
-       ORDER BY mc.name`,
-        [yearStartIso],
-      );
-
-      const totalCurrent = classStats.reduce((s, r) => s + r.total, 0);
-
-      // Section 1: General member status counts
-      const [statusCounts] = await tenantQuery(
-        slug,
-        `SELECT
-         COUNT(*) FILTER (
-           WHERE status_id IN (SELECT id FROM member_statuses WHERE name ILIKE '%Current%')
-             AND (next_renewal IS NULL OR next_renewal < $1::date)
-         )::int AS current_not_renewed,
-         COUNT(*) FILTER (
-           WHERE status_id IN (SELECT id FROM member_statuses WHERE name ILIKE '%Lapsed%')
-         )::int AS lapsed_count
-       FROM members`,
-        [yearStartIso],
-      );
-
-      // Section 3: Active groups
-      const [groupStats] = await tenantQuery(
-        slug,
-        `SELECT
-         COUNT(*)::int AS active_groups,
-         COALESCE(AVG(gm_counts.cnt), 0)::numeric(10,1) AS avg_members
-       FROM groups g
-       LEFT JOIN (
-         SELECT group_id, COUNT(*)::int AS cnt
-         FROM group_members
-         WHERE waiting_since IS NULL
-         GROUP BY group_id
-       ) gm_counts ON gm_counts.group_id = g.id
-       WHERE g.status = 'active'`,
-      );
-
-      const [notInGroup] = await tenantQuery(
-        slug,
-        `SELECT COUNT(*)::int AS count
-       FROM members m
-       WHERE status_id IN (SELECT id FROM member_statuses WHERE name ILIKE '%Current%')
-         AND NOT EXISTS (
-           SELECT 1 FROM group_members gm
-           WHERE gm.member_id = m.id AND gm.waiting_since IS NULL
-         )`,
-      );
-
-      // Section 4: Members by Renew Date
-      const toDate = req.query.to || now.toISOString().slice(0, 10);
-      const fromDate = req.query.from || yearStartIso;
-
-      const renewStats = await tenantQuery(
-        slug,
-        `SELECT mc.id, mc.name,
-              COUNT(m.id) FILTER (
-                WHERE m.status_id IN (SELECT id FROM member_statuses WHERE name ILIKE '%Current%')
-                  AND m.next_renewal IS NOT NULL
-                  AND m.next_renewal >= $1::date
-                  AND m.next_renewal <= $2::date
-              )::int AS not_renewed,
-              COUNT(m.id) FILTER (
-                WHERE m.joined_on >= $1::date
-                  AND m.joined_on <= $2::date
-              )::int AS new_members
-       FROM member_classes mc
-       LEFT JOIN members m ON m.class_id = mc.id
-       GROUP BY mc.id, mc.name
-       ORDER BY mc.name`,
-        [fromDate, toDate],
-      );
-
-      res.json({
-        yearStart: yearStartIso,
-        advanceRenewalsWeeks: advanceWeeks,
-        graceLapseWeeks: graceLapse,
-        currentNotRenewed: statusCounts?.current_not_renewed ?? 0,
-        lapsedCount: statusCounts?.lapsed_count ?? 0,
-        totalCurrent,
-        classStats,
-        activeGroups: groupStats?.active_groups ?? 0,
-        avgGroupMembers: groupStats?.avg_members ?? 0,
-        membersNotInGroup: notInGroup?.count ?? 0,
-        renewStats,
-        renewFrom: fromDate,
-        renewTo: toDate,
-      });
-    } catch (err) {
-      next(err);
-    }
-  },
-);
-
-// ─── GET /members/renewals ────────────────────────────────────────────────
-// Lists Current and Lapsed members with fee info for the renewals screen.
-// Also returns year boundaries so the client can filter by period.
-
-router.get(
-  '/renewals',
-  requireFeature('membershipRenewals'),
-  requirePrivilege('membership_renewals', 'view'),
-  async (req, res, next) => {
-    try {
-      const slug = req.user.tenantSlug;
-
-      const [cfg] = await tenantQuery(
-        slug,
-        `SELECT year_start_month, year_start_day, advance_renewals_weeks
-       FROM tenant_settings WHERE id = 'singleton'`,
-      );
-      const ysm = cfg?.year_start_month ?? 1;
-      const ysd = cfg?.year_start_day ?? 1;
-      const advW = cfg?.advance_renewals_weeks ?? 4;
-
-      const now = new Date();
-      const yr = now.getFullYear();
-      const candidateStart = new Date(yr, ysm - 1, ysd);
-      const yearStart = candidateStart <= now ? candidateStart : new Date(yr - 1, ysm - 1, ysd);
-
-      const prevYearStart = new Date(yearStart);
-      prevYearStart.setFullYear(prevYearStart.getFullYear() - 1);
-
-      const nextYearStart = new Date(yearStart);
-      nextYearStart.setFullYear(nextYearStart.getFullYear() + 1);
-
-      // "Next year" tab only visible within advance_renewals_weeks before nextYearStart
-      const advanceStart = new Date(nextYearStart);
-      advanceStart.setDate(advanceStart.getDate() - advW * 7);
-      const showNextYear = now >= advanceStart;
-
-      const rows = await tenantQuery(
-        slug,
-        `SELECT m.id, m.membership_number, m.forenames, m.surname, m.known_as,
-              ms.id   AS status_id, ms.name AS status_name,
-              mc.id   AS class_id,  mc.name AS class_name,
-              mc.fee, mc.gift_aid_fee,
-              m.next_renewal, m.gift_aid_from, m.partner_id, m.email,
-              m.portal_password_hash IS NOT NULL AS has_portal_password,
-              m.portal_email_verified,
-              p.forenames AS partner_forenames, p.surname AS partner_surname
-       FROM members m
-       LEFT JOIN member_statuses ms ON ms.id = m.status_id
-       LEFT JOIN member_classes  mc ON mc.id = m.class_id
-       LEFT JOIN members          p ON p.id  = m.partner_id
-       WHERE ms.name ILIKE '%Current%' OR ms.name ILIKE '%Lapsed%'
-       ORDER BY m.surname, m.forenames`,
-      );
-
-      res.json({
-        members: rows,
-        yearStart: yearStart.toISOString().slice(0, 10),
-        prevYearStart: prevYearStart.toISOString().slice(0, 10),
-        nextYearStart: nextYearStart.toISOString().slice(0, 10),
-        showNextYear,
-      });
-    } catch (err) {
-      next(err);
-    }
-  },
-);
-
-// ─── POST /members/renew ──────────────────────────────────────────────────
-// Bulk-renews the given members: advances next_renewal by 1 year, sets
-// status to Current if Lapsed, creates a finance transaction for each.
-
-const renewSchema = z.object({
-  memberIds: z.array(z.string().min(1)).min(1),
-  accountId: z.string().min(1),
-  paymentMethod: z.string().min(1),
-  amounts: z.record(z.string(), z.number().positive()),
-  giftAidChanges: z.record(z.string(), z.boolean()).optional(),
-  yearStart: z.string().regex(/^\d{4}-\d{2}-\d{2}$/),
-});
-
-router.post(
-  '/renew',
-  requireFeature('membershipRenewals'),
-  requirePrivilege('membership_renewals', 'renew'),
-  async (req, res, next) => {
-    try {
-      const slug = req.user.tenantSlug;
-      const data = renewSchema.parse(req.body);
-
-      // Find the "Current" status id to upgrade Lapsed members
-      const [currentStatus] = await tenantQuery(
-        slug,
-        `SELECT id FROM member_statuses WHERE name ILIKE '%Current%' ORDER BY name LIMIT 1`,
-      );
-      if (!currentStatus) throw new AppError('No "Current" member status found.', 400);
-
-      const renewed = [];
-      const errors = [];
-
-      for (const memberId of data.memberIds) {
-        try {
-          // Fetch member to get current next_renewal and status
-          const [m] = await tenantQuery(
-            slug,
-            `SELECT m.id, m.forenames, m.surname, m.next_renewal, m.status_id,
-                  ms.name AS status_name, m.gift_aid_from, m.class_id
-           FROM members m
-           LEFT JOIN member_statuses ms ON ms.id = m.status_id
-           WHERE m.id = $1`,
-            [memberId],
-          );
-          if (!m) {
-            errors.push({ memberId, error: 'Member not found' });
-            continue;
-          }
-
-          // New next_renewal = current next_renewal + 1 year (or yearStart + 1 year if null)
-          const base = m.next_renewal ? String(m.next_renewal).slice(0, 10) : data.yearStart;
-          const baseDate = new Date(base);
-          baseDate.setFullYear(baseDate.getFullYear() + 1);
-          const newNextRenewal = baseDate.toISOString().slice(0, 10);
-
-          // Should we set to Current?
-          const isLapsed = (m.status_name ?? '').toLowerCase().includes('lapsed');
-          const newStatusId = isLapsed ? currentStatus.id : m.status_id;
-
-          // Gift Aid update
-          const gaChange = data.giftAidChanges?.[memberId];
-          let giftAidFrom;
-          if (gaChange === true && !m.gift_aid_from) {
-            giftAidFrom = 'TODAY';
-          } else if (gaChange === false) {
-            giftAidFrom = null;
-          } else {
-            giftAidFrom = m.gift_aid_from ? String(m.gift_aid_from).slice(0, 10) : null;
-          }
-
-          // Update member record
-          await tenantQuery(
-            slug,
-            `UPDATE members
-           SET next_renewal  = $1::date,
-               status_id     = $2,
-               gift_aid_from = $3::date,
-               card_printed  = false,
-               updated_at    = now()
-           WHERE id = $4`,
-            [
-              newNextRenewal,
-              newStatusId,
-              giftAidFrom === 'TODAY' ? new Date().toISOString().slice(0, 10) : giftAidFrom,
-              memberId,
-            ],
-          );
-
-          // Create finance transaction (no category splits — user can categorize via ledger)
-          const amount = data.amounts[memberId];
-          const fromTo = `${m.forenames} ${m.surname}`;
-
-          // Resolve Gift Aid eligible amount (uses the member's gift_aid_from after any update above)
-          const effectiveGaFrom =
-            giftAidFrom === 'TODAY' ? new Date().toISOString().slice(0, 10) : giftAidFrom;
-          let gaAmount = null;
-          if (effectiveGaFrom && m.class_id) {
-            gaAmount = await resolveGiftAidAmount(
-              slug,
-              memberId,
-              m.class_id,
-              new Date().toISOString().slice(0, 10),
-            );
-          }
-
-          const [txn] = await tenantQuery(
-            slug,
-            `INSERT INTO transactions
-             (account_id, date, type, from_to, amount, payment_method, detail, member_id_1, gift_aid_amount)
-           VALUES ($1, CURRENT_DATE, 'in', $2, $3::numeric, $4, 'Membership', $5, $6::numeric)
-           RETURNING id, transaction_number`,
-            [data.accountId, fromTo, amount, data.paymentMethod, memberId, gaAmount],
-          );
-
-          logAudit(slug, {
-            userId: req.user.userId,
-            userName: req.user.name,
-            action: 'renew',
-            entityType: 'member',
-            entityId: memberId,
-            detail: JSON.stringify({ newNextRenewal }),
-          });
-
-          // Gift Aid consent audit
-          if (gaChange === true && !m.gift_aid_from) {
-            logAudit(slug, {
-              userId: req.user.userId,
-              userName: req.user.name,
-              action: 'gift_aid_consent',
-              entityType: 'member',
-              entityId: memberId,
-              entityName: `${m.forenames} ${m.surname}`,
-              detail: JSON.stringify({ giftAidFrom: new Date().toISOString().slice(0, 10) }),
-            });
-          } else if (gaChange === false && m.gift_aid_from) {
-            logAudit(slug, {
-              userId: req.user.userId,
-              userName: req.user.name,
-              action: 'gift_aid_withdrawn',
-              entityType: 'member',
-              entityId: memberId,
-              entityName: `${m.forenames} ${m.surname}`,
-              detail: JSON.stringify({ previousGiftAidFrom: String(m.gift_aid_from).slice(0, 10) }),
-            });
-          }
-
-          renewed.push({ memberId, newNextRenewal, transactionNumber: txn.transaction_number });
-        } catch (err) {
-          errors.push({ memberId, error: err.message });
-        }
-      }
-
-      res.json({ renewed, errors });
-    } catch (err) {
-      next(err);
-    }
-  },
-);
-
-// ─── GET /members/non-renewals ────────────────────────────────────────────
-// mode=this_year  — Current members whose next_renewal < current year start
-// mode=long_term  — All members whose next_renewal is older than deletion_years
-
-router.get(
-  '/non-renewals',
-  requireFeature('membershipRenewals'),
-  requirePrivilege('members_non_renewals', 'view'),
-  async (req, res, next) => {
-    try {
-      const slug = req.user.tenantSlug;
-      const mode = req.query.mode === 'long_term' ? 'long_term' : 'this_year';
-
-      const [cfg] = await tenantQuery(
-        slug,
-        `SELECT year_start_month, year_start_day, grace_lapse_weeks, deletion_years
-       FROM tenant_settings WHERE id = 'singleton'`,
-      );
-      const ysm = cfg?.year_start_month ?? 1;
-      const ysd = cfg?.year_start_day ?? 1;
-      const graceLapse = cfg?.grace_lapse_weeks ?? 4;
-      const deletionYrs = cfg?.deletion_years ?? 7;
-
-      const now = new Date();
-      const yr = now.getFullYear();
-      const candidateStart = new Date(yr, ysm - 1, ysd);
-      const yearStart = candidateStart <= now ? candidateStart : new Date(yr - 1, ysm - 1, ysd);
-      const yearStartIso = yearStart.toISOString().slice(0, 10);
-
-      let rows;
-      if (mode === 'this_year') {
-        rows = await tenantQuery(
-          slug,
-          `SELECT m.id, m.membership_number, m.forenames, m.surname, m.known_as,
-                ms.name AS status_name,
-                mc.name AS class_name,
-                m.next_renewal, m.email, m.mobile,
-                a.house_no, a.street, a.town, a.postcode, a.telephone,
-                m.portal_password_hash IS NOT NULL AS has_portal_password,
-                m.portal_email_verified,
-                (SELECT EXTRACT(YEAR FROM MAX(t.date))::int
-                 FROM transactions t
-                 WHERE t.member_id_1 = m.id AND t.type = 'in') AS last_renewal_year
-         FROM members m
-         LEFT JOIN member_statuses ms ON ms.id = m.status_id
-         LEFT JOIN member_classes  mc ON mc.id = m.class_id
-         LEFT JOIN addresses        a ON a.id  = m.address_id
-         WHERE ms.name ILIKE '%Current%'
-           AND (m.next_renewal IS NULL OR m.next_renewal < $1::date)
-         ORDER BY m.surname, m.forenames`,
-          [yearStartIso],
-        );
-      } else {
-        // Compute cutoff date in JS to avoid PostgreSQL interval casting issues
-        const cutoff = new Date();
-        cutoff.setFullYear(cutoff.getFullYear() - deletionYrs);
-        const cutoffIso = cutoff.toISOString().slice(0, 10);
-
-        rows = await tenantQuery(
-          slug,
-          `SELECT m.id, m.membership_number, m.forenames, m.surname, m.known_as,
-                ms.name AS status_name,
-                mc.name AS class_name,
-                m.next_renewal, m.email, m.mobile,
-                a.house_no, a.street, a.town, a.postcode, a.telephone,
-                m.portal_password_hash IS NOT NULL AS has_portal_password,
-                m.portal_email_verified,
-                (SELECT EXTRACT(YEAR FROM MAX(t.date))::int
-                 FROM transactions t
-                 WHERE t.member_id_1 = m.id AND t.type = 'in') AS last_renewal_year
-         FROM members m
-         LEFT JOIN member_statuses ms ON ms.id = m.status_id
-         LEFT JOIN member_classes  mc ON mc.id = m.class_id
-         LEFT JOIN addresses        a ON a.id  = m.address_id
-         WHERE m.next_renewal IS NOT NULL
-           AND m.next_renewal < $1::date
-         ORDER BY m.next_renewal, m.surname`,
-          [cutoffIso],
-        );
-      }
-
-      res.json({
-        members: rows,
-        mode,
-        yearStart: yearStartIso,
-        graceLapse,
-        deletionYears: deletionYrs,
-      });
-    } catch (err) {
-      next(err);
-    }
-  },
-);
-
-// ─── POST /members/lapse ──────────────────────────────────────────────────
-// Changes status to the "Lapsed" status for the given member IDs.
-
-router.post(
-  '/lapse',
-  requireFeature('membershipRenewals'),
-  requirePrivilege('members_non_renewals', 'lapse'),
-  async (req, res, next) => {
-    try {
-      const slug = req.user.tenantSlug;
-      const { memberIds } = z.object({ memberIds: z.array(z.string()).min(1) }).parse(req.body);
-
-      const [lapsedStatus] = await tenantQuery(
-        slug,
-        `SELECT id FROM member_statuses WHERE name ILIKE '%Lapsed%' ORDER BY name LIMIT 1`,
-      );
-      if (!lapsedStatus) throw new AppError('No "Lapsed" member status found.', 400);
-
-      await tenantQuery(
-        slug,
-        `UPDATE members SET status_id = $1, updated_at = now()
-       WHERE id = ANY($2::text[])`,
-        [lapsedStatus.id, memberIds],
-      );
-
-      logAudit(slug, {
-        userId: req.user.userId,
-        userName: req.user.name,
-        action: 'lapse',
-        entityType: 'member',
-        entityId: memberIds.join(','),
-      });
-      res.json({ lapsed: memberIds.length });
-    } catch (err) {
-      next(err);
-    }
-  },
-);
-
-router.get(
-  '/validate',
-  requirePrivilege('member_data_validation', 'view'),
-  async (req, res, next) => {
-    try {
-      const members = await tenantQuery(
-        req.user.tenantSlug,
-        `SELECT m.id, m.membership_number, m.forenames, m.surname,
-              m.status_id, m.class_id, m.joined_on, m.next_renewal,
-              m.email, m.mobile,
-              a.id         AS address_id,
-              a.house_no, a.street, a.add_line1, a.add_line2,
-              a.town, a.county, a.postcode, a.telephone
-       FROM members m
-       LEFT JOIN addresses a ON a.id = m.address_id
-       ORDER BY m.surname, m.forenames`,
-      );
-      res.json(members);
-    } catch (err) {
-      next(err);
-    }
-  },
-);
-
-// ─── GET /members/download ────────────────────────────────────────────────
-// Download selected members as Excel, PDF, or email-CSV.
-// Query params: format (excel|pdf|email-csv), ids (comma-separated), fields (comma-separated)
-
-const MEMBER_FIELD_DEFS = {
-  membership_number: { label: 'Membership No', get: (m) => String(m.membership_number ?? '') },
-  title: { label: 'Title', get: (m) => m.title ?? '' },
-  forenames: { label: 'Forenames', get: (m) => m.forenames ?? '' },
-  known_as: { label: 'Known As', get: (m) => m.known_as ?? '' },
-  surname: { label: 'Surname', get: (m) => m.surname ?? '' },
-  email: { label: 'Email', get: (m) => m.email ?? '' },
-  mobile: { label: 'Mobile', get: (m) => m.mobile ?? '' },
-  telephone: { label: 'Telephone', get: (m) => m.telephone ?? '' },
-  address: {
-    label: 'Address',
-    get: (m) => [m.house_no, m.street, m.add_line1, m.add_line2].filter(Boolean).join(', '),
-  },
-  town: { label: 'Town', get: (m) => m.town ?? '' },
-  county: { label: 'County', get: (m) => m.county ?? '' },
-  postcode: { label: 'Postcode', get: (m) => m.postcode ?? '' },
-  country: { label: 'Country', get: (m) => m.country ?? '' },
-  status: { label: 'Status', get: (m) => m.status ?? '' },
-  class: { label: 'Class', get: (m) => m.class ?? '' },
-  joined_on: { label: 'Joined', get: (m) => (m.joined_on ? String(m.joined_on).slice(0, 10) : '') },
-  next_renewal: {
-    label: 'Next Renewal',
-    get: (m) => (m.next_renewal ? String(m.next_renewal).slice(0, 10) : ''),
-  },
-  custom_field_1: { label: 'Custom Field 1', get: (m) => m.custom_field_1 ?? '' },
-  custom_field_2: { label: 'Custom Field 2', get: (m) => m.custom_field_2 ?? '' },
-  custom_field_3: { label: 'Custom Field 3', get: (m) => m.custom_field_3 ?? '' },
-  custom_field_4: { label: 'Custom Field 4', get: (m) => m.custom_field_4 ?? '' },
-  emergency_contact: { label: 'Emergency Contact', get: (m) => m.emergency_contact ?? '' },
-};
-
-function buildMemberPdf(rows, cols, title) {
-  const PAGE_W = 841.89;
-  const PAGE_H = 595.28; // A4 landscape
-  const MARGIN = 36;
-  const FONT_SZ = 7;
-  const ROW_H = 13;
-  const usableW = PAGE_W - MARGIN * 2;
-  const colW = usableW / cols.length;
-
-  const doc = new PDFDocument({
-    margin: MARGIN,
-    size: 'A4',
-    layout: 'landscape',
-    autoFirstPage: true,
-  });
-
-  function drawHeader(y) {
-    doc.font('Helvetica-Bold').fontSize(FONT_SZ);
-    cols.forEach((f, idx) => {
-      doc.text(MEMBER_FIELD_DEFS[f].label, MARGIN + idx * colW, y, {
-        width: colW - 3,
-        lineBreak: false,
-        ellipsis: true,
-      });
-    });
-    return y + ROW_H;
-  }
-
-  let y = MARGIN + 4;
-  doc.font('Helvetica-Bold').fontSize(9).text(title, MARGIN, y, { lineBreak: false });
-  y += 16;
-  y = drawHeader(y);
-  doc
-    .moveTo(MARGIN, y - 2)
-    .lineTo(PAGE_W - MARGIN, y - 2)
-    .strokeColor('#aaaaaa')
-    .stroke();
-
-  doc.font('Helvetica').fontSize(FONT_SZ);
-  for (const m of rows) {
-    if (y + ROW_H > PAGE_H - MARGIN) {
-      doc.addPage({ size: 'A4', layout: 'landscape' });
-      y = MARGIN + 4;
-      y = drawHeader(y);
-      doc
-        .moveTo(MARGIN, y - 2)
-        .lineTo(PAGE_W - MARGIN, y - 2)
-        .strokeColor('#aaaaaa')
-        .stroke();
-      doc.font('Helvetica').fontSize(FONT_SZ);
-    }
-    cols.forEach((f, idx) => {
-      doc.text(MEMBER_FIELD_DEFS[f].get(m), MARGIN + idx * colW, y, {
-        width: colW - 3,
-        lineBreak: false,
-        ellipsis: true,
-      });
-    });
-    y += ROW_H;
-  }
-  doc.end();
-  return doc;
-}
-
-router.get('/download', requirePrivilege('members_list', 'view'), async (req, res, next) => {
-  try {
-    const { format = 'excel', ids = '', fields = '' } = req.query;
-    const slug = req.user.tenantSlug;
-    const memberIds = ids
-      .split(',')
-      .map((s) => s.trim())
-      .filter(Boolean);
-    if (!memberIds.length) throw AppError('No member IDs provided.', 400);
-
-    const rows = await tenantQuery(
-      slug,
-      `SELECT m.id, m.membership_number, m.title, m.forenames, m.surname, m.known_as,
-              m.email, m.mobile,
-              ms.name AS status, mc.name AS class,
-              a.house_no, a.street, a.add_line1, a.add_line2,
-              a.town, a.county, a.postcode, a.country, a.telephone,
-              m.joined_on, m.next_renewal,
-              m.custom_field_1, m.custom_field_2,
-              m.custom_field_3, m.custom_field_4
-       FROM members m
-       LEFT JOIN member_statuses ms ON ms.id = m.status_id
-       LEFT JOIN member_classes  mc ON mc.id = m.class_id
-       LEFT JOIN addresses        a ON a.id  = m.address_id
-       WHERE m.id = ANY($1::text[])
-       ORDER BY m.surname, m.forenames`,
-      [memberIds],
-    );
-
-    const tenantPart = slug.replace(/^u3a_/, '');
-    const stamp = new Date().toISOString().slice(0, 10);
-
-    if (format === 'email-csv') {
-      const content = rows
-        .map((m) => m.email)
-        .filter(Boolean)
-        .join('\n');
-      res.setHeader('Content-Type', 'text/plain; charset=utf-8');
-      res.setHeader(
-        'Content-Disposition',
-        `attachment; filename="${tenantPart}_member_emails_${stamp}.csv"`,
-      );
-      return res.send(content);
-    }
-
-    const activeCols = fields
-      .split(',')
-      .map((s) => s.trim())
-      .filter((f) => f && MEMBER_FIELD_DEFS[f]);
-    const cols = activeCols.length ? activeCols : Object.keys(MEMBER_FIELD_DEFS);
-
-    if (format === 'excel') {
-      const wb = new ExcelJS.Workbook();
-      const ws = wb.addWorksheet('Members');
-      ws.columns = cols.map((f) => ({ header: MEMBER_FIELD_DEFS[f].label, width: 20 }));
-      ws.getRow(1).font = { bold: true };
-      for (const m of rows) ws.addRow(cols.map((f) => sanitizeCell(MEMBER_FIELD_DEFS[f].get(m))));
-      res.setHeader(
-        'Content-Type',
-        'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
-      );
-      res.setHeader(
-        'Content-Disposition',
-        `attachment; filename="${tenantPart}_members_${stamp}.xlsx"`,
-      );
-      await wb.xlsx.write(res);
-      return res.end();
-    }
-
-    if (format === 'pdf') {
-      const doc = buildMemberPdf(rows, cols, `Members — ${stamp}`);
-      const chunks = [];
-      doc.on('data', (c) => chunks.push(c));
-      await new Promise((resolve) => doc.on('end', resolve));
-      const buf = Buffer.concat(chunks);
-      res.setHeader('Content-Type', 'application/pdf');
-      res.setHeader(
-        'Content-Disposition',
-        `attachment; filename="${tenantPart}_members_${stamp}.pdf"`,
-      );
-      return res.send(buf);
-    }
-
-    throw AppError('Invalid format.', 400);
-  } catch (err) {
-    next(err);
-  }
-});
 
 // ─── GET /members/:id/groups ──────────────────────────────────────────────
 // Returns groups (and waiting list) the member belongs to.
